KH05802CLI-65 (#8)
